+++ tests/tests_unit/test_api/test_assets.py
@@ -253,7 +253,7 @@ class TestAssetPoster:
 
     def test_validate_asset_hierarchy__more_than_limit_only_resolved_assets(self):
         with set_request_limit(ASSETS_API, 1):
-            _AssetPoster([Asset(parent_id=1), Asset(parent_id=2)], ASSETS_API)
+            _AssetPoster([Asset(external_id="a1", parent_id=1), Asset(external_id="a2", parent_id=2)], ASSETS_API)
 
     def test_validate_asset_hierarchy_circular_dependencies(self):
         assets = [
@@ -358,7 +358,7 @@ class TestAssetPoster:
     def test_post_hierarchy(self, limit, depth, children_per_node, expected_num_calls, mock_post_asset_hierarchy):
         assets = generate_asset_tree(root_external_id="0", depth=depth, children_per_node=children_per_node)
         with set_request_limit(ASSETS_API, limit):
-            created_assets = ASSETS_API.create(assets)
+            created_assets = ASSETS_API.create_hierarchy(assets)
 
         assert len(assets) == len(created_assets)
         assert expected_num_calls - 1 <= len(mock_post_asset_hierarchy.calls) <= expected_num_calls + 1
@@ -370,7 +370,9 @@ class TestAssetPoster:
 
     def test_post_assets_over_limit_only_resolved(self, mock_post_asset_hierarchy):
         with set_request_limit(ASSETS_API, 1):
-            _AssetPoster([Asset(parent_id=1), Asset(parent_id=2)], ASSETS_API).post()
+            _AssetPoster(
+                [Asset(external_id="a1", parent_id=1), Asset(external_id="a2", parent_id=2)], ASSETS_API
+            ).post()
         assert 2 == len(mock_post_asset_hierarchy.calls)
 
     @pytest.fixture
@@ -422,7 +424,7 @@ class TestAssetPoster:
             Asset(name="200", parent_external_id="03", external_id="031"),
         ]
         with pytest.raises(CogniteAPIError) as e:
-            ASSETS_API.create(assets)
+            ASSETS_API.create_hierarchy(assets)
 
         assert {a.external_id for a in e.value.unknown} == {"03"}
         assert {a.external_id for a in e.value.failed} == {"02", "021", "0211", "031"}