+++ tests/tests_unit/test_testing.py
@@ -1,0 +1,31 @@
+from unittest.mock import MagicMock
+
+import pytest
+
+from cognite.client import CogniteClient
+from cognite.client.testing import mock_cognite_client
+
+
+def test_mock_cognite_client():
+    with mock_cognite_client() as c_mock:
+        c = CogniteClient()
+        assert isinstance(c_mock, MagicMock)
+        assert c_mock == c
+
+        api_pairs = [
+            (c.time_series, c_mock.time_series),
+            (c.raw, c_mock.raw),
+            (c.assets, c_mock.assets),
+            (c.datapoints, c_mock.datapoints),
+            (c.events, c_mock.events),
+            (c.files, c_mock.files),
+            (c.iam, c_mock.iam),
+            (c.login, c_mock.login),
+            (c.three_d, c_mock.three_d),
+        ]
+        for api, mock_api in api_pairs:
+            assert isinstance(mock_api, MagicMock)
+            assert api == mock_api
+
+            with pytest.raises(AttributeError):
+                api.does_not_exist