+++ CHANGELOG.md
@@ -21,6 +21,20 @@ Changes are grouped as follows
 
 ## [Unreleased]
 
+## [1.1.0] - 2019-08-21
+### Added
+- New method create_hierarchy() added to assets API.
+- SequencesAPI.list now accepts an asset_ids parameter for searching by asset
+- SequencesDataAPI.insert now accepts a SequenceData object for easier copying
+- DatapointsAPI.insert now accepts a Datapoints object for easier copying
+- helper method `cognite.client.testing.mock_cognite_client()` for mocking CogniteClient
+- parent_id and parent_external_id to AssetUpdate class.
+- source_created_time and source_modified_time support for files
+
+### Changed
+- assets.create() no longer validates asset hierarchy and sorts assets before posting. This functionality has been moved to assets.create_hierarchy().
+- AssetList.files() and AssetList.events() now deduplicate results while fetching related resources, significantly reducing memory load.
+
 ## [1.0.5] - 2019-08-15
 ### Added
 - files.create() method to enable creating a file without uploading content.